Added book inspect system
Modified the current note inspection system to differentiate between normal notes and books.
