+++ .vs/Enigmatika_Version_Control/v17/DocumentLayout.json
@@ -3,16 +3,12 @@
   "WorkspaceRootPath": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\",
   "Documents": [
     {
-      "AbsoluteMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\assets\\noteinspectionmanager.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
-      "RelativeMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|solutionrelative:assets\\noteinspectionmanager.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
-    },
-    {
-      "AbsoluteMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|c:\\unity game projects (jj)\\enigmatika_version_control\\assets\\noteobjecthandler.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
-      "RelativeMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|solutionrelative:assets\\noteobjecthandler.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
-    },
-    {
-      "AbsoluteMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|c:\\unity game projects (jj)\\enigmatika_version_control\\assets\\ui\\ui elements\\scripts\\backbuttonhandler.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "AbsoluteMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\assets\\ui\\ui elements\\scripts\\backbuttonhandler.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
       "RelativeMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|solutionrelative:assets\\ui\\ui elements\\scripts\\backbuttonhandler.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|c:\\unity game projects (jj)\\enigmatika_version_control\\assets\\noteuicontroller.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{FE007D7E-777E-F10D-903C-F74F7F6AC41E}|Assembly-CSharp.csproj|solutionrelative:assets\\noteuicontroller.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
     }
   ],
   "DocumentGroupContainers": [
@@ -22,45 +18,32 @@
       "DocumentGroups": [
         {
           "DockedWidth": 200,
-          "SelectedChildIndex": 2,
+          "SelectedChildIndex": 1,
           "Children": [
             {
               "$type": "Document",
               "DocumentIndex": 1,
-              "Title": "NoteObjectHandler.cs",
-              "DocumentMoniker": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\NoteObjectHandler.cs",
-              "RelativeDocumentMoniker": "Assets\\NoteObjectHandler.cs",
-              "ToolTip": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\NoteObjectHandler.cs",
-              "RelativeToolTip": "Assets\\NoteObjectHandler.cs",
-              "ViewState": "AgIAAAAAAAAAAAAAAAAAABMAAAAAAAAAAAAAAA==",
+              "Title": "NoteUIController.cs",
+              "DocumentMoniker": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\NoteUIController.cs",
+              "RelativeDocumentMoniker": "Assets\\NoteUIController.cs",
+              "ToolTip": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\NoteUIController.cs",
+              "RelativeToolTip": "Assets\\NoteUIController.cs",
+              "ViewState": "AgIAAAMAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
-              "WhenOpened": "2024-10-01T11:59:41.69Z",
-              "EditorCaption": ""
-            },
-            {
-              "$type": "Document",
-              "DocumentIndex": 2,
-              "Title": "BackButtonHandler.cs",
-              "DocumentMoniker": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
-              "RelativeDocumentMoniker": "Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
-              "ToolTip": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
-              "RelativeToolTip": "Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
-              "ViewState": "AgIAAD0AAAAAAAAAAAAAAFMAAAAAAAAAAAAAAA==",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
-              "WhenOpened": "2024-10-01T15:21:28.515Z",
+              "WhenOpened": "2024-10-04T03:12:32.763Z",
               "EditorCaption": ""
             },
             {
               "$type": "Document",
               "DocumentIndex": 0,
-              "Title": "NoteInspectionManager.cs",
-              "DocumentMoniker": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\NoteInspectionManager.cs",
-              "RelativeDocumentMoniker": "Assets\\NoteInspectionManager.cs",
-              "ToolTip": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\NoteInspectionManager.cs",
-              "RelativeToolTip": "Assets\\NoteInspectionManager.cs",
-              "ViewState": "AgIAALIAAAAAAAAAAAAAANcAAAABAAAAAAAAAA==",
+              "Title": "BackButtonHandler.cs",
+              "DocumentMoniker": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
+              "RelativeDocumentMoniker": "Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
+              "ToolTip": "C:\\Unity game projects (JJ)\\Enigmatika_Version_Control\\Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
+              "RelativeToolTip": "Assets\\UI\\UI elements\\SCRIPTS\\BackButtonHandler.cs",
+              "ViewState": "AgIAABwAAAAAAAAAAAAAwFcAAAAAAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
-              "WhenOpened": "2024-10-01T05:40:01.574Z",
+              "WhenOpened": "2024-10-01T15:21:28.515Z",
               "EditorCaption": ""
             }
           ]
