+++ Assets/Interaction system/Interactable Objects/Note Inspection/NoteUIController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c75d6aa212d79a6499416fe0993f880a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
