--- Assets/NoteUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c75d6aa212d79a6499416fe0993f880a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
