+++ Assets/Test assets/Books.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8f67ad033f25be44da57d9f8ea3544f4
+folderAsset: yes
+timeCreated: 1520654001
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
