+++ Assets/Test assets/Books/3dobjects_FBX.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 57ec6b8ecc8061349a1050c87f9e7505
+folderAsset: yes
+timeCreated: 1520654001
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
