+++ Assets/Test assets/Books/3dobjects_FBX/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 77b29326df09434499ce1c173ca26046
+folderAsset: yes
+timeCreated: 1520654001
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
