+++ Assets/Test assets/Books/3dobjects_FBX/Materials/books-book.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3d87070dc9fb4246b3a788982ab59e8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
