+++ Assets/Test assets/Books/3dobjects_FBX/Materials/books.mat.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 153061f29d41be9438999ac2fae2c894
+NativeFormatImporter:
+  mainObjectFileID: -1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
