+++ Assets/Test assets/Books/Prefabs.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: efff46687f43f8144a5cd9e509b5280b
+folderAsset: yes
+timeCreated: 1520654001
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
