+++ Assets/Test assets/Books/Prefabs/book_0001c.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 888ce82c0af2ab0499d39c93fb0f4ce4
+NativeFormatImporter:
+  mainObjectFileID: -1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
