+++ Assets/Test assets/Books/Prefabs/book_0001d.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f99df1970d4b71441a320ce6acc52709
+NativeFormatImporter:
+  mainObjectFileID: -1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
