+++ Assets/Test assets/Books/Textures.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9af5ed91f96a875428a6fd76e9ea953c
+folderAsset: yes
+timeCreated: 1520654001
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
