+++ Assets/UI/Reading UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 08a6fcdfcc1cae24c951d423aa912d4e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
