+++ Assets/UI/Reading UI/Books.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82cd277552fae1e49a6fb7f7816cb4a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
