+++ Assets/UI/Reading UI/Notes and Letters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a25cf163709c64b43ad62a472f093cf6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
